Rename LeftNav files and fix imports
