+++ src/components/LeftNav/LeftNav.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classnames from 'classnames';
 
-import styles from './index.module.scss';
+import styles from './LeftNav.module.scss';
 
 const LeftNav = ({ className, children }) => <nav className={classnames(styles.LeftNav, className)}>{children}</nav>;
